[rename] distributeLock관련 클래스 패키지 변경 / 테스트코드 위치 변경
